// frontend/lib/api/generated/model/cronogramaEntregaIn.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */

export interface CronogramaEntregaIn {
  /** @maxLength 20 */
  periodo: string;
  fecha_programada: string;
}